requirements.txt
streamlit
ultralytics
opencv-python-headless
google-generativeai
gtts
Pillow
numpy